refactor: deployment actions
