+++ app/src/components/forms/DeploymentActions.vue
@@ -1,0 +1,423 @@
+<template>
+  <div class="flex flex-col gap-4 mt-4">
+    <div class="flex flex-wrap gap-2">
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="!isBankDeployed && !isLoadingDeployBank && !isConfirmingDeployBank"
+        @click="deployBankAccount"
+      >
+        Deploy Bank
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployBank || isConfirmingDeployBank"
+      />
+
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="!isVotingDeployed && !isLoadingDeployVoting && !isConfirmingDeployVoting"
+        @click="deployVotingContract"
+      >
+        Deploy Voting
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployVoting || isConfirmingDeployVoting"
+      />
+
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="!isExpenseDeployed && !isLoadingDeployExpense && !isConfirmingDeployExpense"
+        @click="deployExpenseAccount"
+      >
+        Deploy Expense
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployExpense || isConfirmingDeployExpense"
+      />
+
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="
+          !isExpenseEip712Deployed &&
+          !isLoadingDeployExpenseEip712 &&
+          !isConfirmingDeployExpenseEip712
+        "
+        @click="deployExpenseAccountEip712"
+      >
+        Deploy Expense EIP712
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployExpenseEip712 || isConfirmingDeployExpenseEip712"
+      />
+
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="!isBoDDeployed && !isLoadingDeployBoD && !isConfirmingDeployBoD"
+        @click="deployBoardOfDirectors"
+      >
+        Deploy Board of Directors
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployBoD || isConfirmingDeployBoD"
+      />
+    </div>
+
+    <div class="flex justify-center">
+      <button
+        class="btn btn-primary btn-sm"
+        v-if="!isLoadingDeployAll && !areAllContractsDeployed && !isConfirmingDeployAll"
+        @click="deployAllContracts"
+      >
+        Deploy All Contracts
+      </button>
+      <LoadingButton
+        :color="'primary min-w-24'"
+        v-if="isLoadingDeployAll || isConfirmingDeployAll"
+      />
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { computed, watch } from 'vue'
+import { useToastStore } from '@/stores'
+import { useUserDataStore } from '@/stores/user'
+import { useWriteContract, useWaitForTransactionReceipt } from '@wagmi/vue'
+import { encodeFunctionData, type Address } from 'viem'
+import { log, parseError } from '@/utils'
+import LoadingButton from '@/components/LoadingButton.vue'
+
+// Contract ABIs and constants
+import OfficerABI from '@/artifacts/abi/officer.json'
+import BankABI from '@/artifacts/abi/Bank.json'
+import VotingABI from '@/artifacts/abi/Voting.json'
+import ExpenseAccountABI from '@/artifacts/abi/expense-account.json'
+import ExpenseAccountEIP712ABI from '@/artifacts/abi/expense-account-eip712.json'
+import BoardOfDirectorsABI from '@/artifacts/abi/bod.json'
+import { TIPS_ADDRESS } from '@/constant'
+import type { Team } from '@/types'
+
+const props = defineProps<{
+  team: Team
+  founders: string[]
+  isBankDeployed: boolean
+  isVotingDeployed: boolean
+  isBoDDeployed: boolean
+  isExpenseDeployed: boolean
+  isExpenseEip712Deployed: boolean
+}>()
+
+const emits = defineEmits(['getTeam'])
+const { addErrorToast, addSuccessToast } = useToastStore()
+
+// Bank deployment
+const {
+  writeContract: deployBank,
+  isPending: isLoadingDeployBank,
+  data: deployBankHash,
+  error: deployBankError
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployBank, isSuccess: isConfirmedDeployBank } =
+  useWaitForTransactionReceipt({ hash: deployBankHash })
+
+watch(isConfirmingDeployBank, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployBank.value) {
+    addSuccessToast('Bank deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// Voting deployment
+const {
+  writeContract: deployVoting,
+  isPending: isLoadingDeployVoting,
+  data: deployVotingHash,
+  error: deployVotingError
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployVoting, isSuccess: isConfirmedDeployVoting } =
+  useWaitForTransactionReceipt({ hash: deployVotingHash })
+
+watch(isConfirmingDeployVoting, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployVoting.value) {
+    addSuccessToast('Voting deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// Expense deployment
+const {
+  writeContract: deployExpense,
+  isPending: isLoadingDeployExpense,
+  data: deployExpenseHash,
+  error: deployExpenseError
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployExpense, isSuccess: isConfirmedDeployExpense } =
+  useWaitForTransactionReceipt({ hash: deployExpenseHash })
+
+watch(isConfirmingDeployExpense, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployExpense.value) {
+    addSuccessToast('Expense account deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// Expense EIP712 deployment
+const {
+  writeContract: deployExpenseEip712,
+  isPending: isLoadingDeployExpenseEip712,
+  data: deployExpenseEip712Hash
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployExpenseEip712, isSuccess: isConfirmedDeployExpenseEip712 } =
+  useWaitForTransactionReceipt({ hash: deployExpenseEip712Hash })
+
+watch(isConfirmingDeployExpenseEip712, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployExpenseEip712.value) {
+    addSuccessToast('Expense EIP712 account deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// BoD deployment
+const {
+  writeContract: deployBoD,
+  isPending: isLoadingDeployBoD,
+  data: deployBoDHash,
+  error: deployBoDError
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployBoD, isSuccess: isConfirmedDeployBoD } =
+  useWaitForTransactionReceipt({ hash: deployBoDHash })
+
+watch(isConfirmingDeployBoD, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployBoD.value) {
+    addSuccessToast('Board of Directors deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// Deploy All Contracts
+const {
+  writeContract: deployAll,
+  isPending: isLoadingDeployAll,
+  error: deployAllError,
+  data: deployAllData
+} = useWriteContract()
+
+const { isLoading: isConfirmingDeployAll, isSuccess: isConfirmedDeployAll } =
+  useWaitForTransactionReceipt({
+    hash: deployAllData
+  })
+
+watch(isConfirmingDeployAll, (isConfirming, wasConfirming) => {
+  if (wasConfirming && !isConfirming && isConfirmedDeployAll.value) {
+    addSuccessToast('All contracts deployed successfully')
+    emits('getTeam')
+  }
+})
+
+// Deployment functions
+const deployBankAccount = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const initData = encodeFunctionData({
+    abi: BankABI,
+    functionName: 'initialize',
+    args: [TIPS_ADDRESS, currentAddress]
+  })
+
+  deployBank({
+    address: props.team.officerAddress as Address,
+    abi: OfficerABI,
+    functionName: 'deployBeaconProxy',
+    args: ['Bank', initData]
+  })
+}
+
+const deployVotingContract = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const initData = encodeFunctionData({
+    abi: VotingABI,
+    functionName: 'initialize',
+    args: [currentAddress]
+  })
+
+  deployVoting({
+    address: props.team.officerAddress as Address,
+    abi: OfficerABI,
+    functionName: 'deployBeaconProxy',
+    args: ['Voting', initData]
+  })
+}
+
+const deployExpenseAccount = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const initData = encodeFunctionData({
+    abi: ExpenseAccountABI,
+    functionName: 'initialize',
+    args: [currentAddress]
+  })
+
+  deployExpense({
+    address: props.team.officerAddress as Address,
+    abi: OfficerABI,
+    functionName: 'deployBeaconProxy',
+    args: ['ExpenseAccount', initData]
+  })
+}
+
+const deployExpenseAccountEip712 = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const initData = encodeFunctionData({
+    abi: ExpenseAccountEIP712ABI,
+    functionName: 'initialize',
+    args: [currentAddress]
+  })
+
+  deployExpenseEip712({
+    address: props.team.officerAddress as Address,
+    abi: OfficerABI,
+    functionName: 'deployBeaconProxy',
+    args: ['ExpenseAccountEIP712', initData]
+  })
+}
+
+const deployBoardOfDirectors = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const owners = [...props.founders, currentAddress]
+  const initData = encodeFunctionData({
+    abi: BoardOfDirectorsABI,
+    functionName: 'initialize',
+    args: [owners]
+  })
+
+  deployBoD({
+    address: props.team.officerAddress as Address,
+    abi: OfficerABI,
+    functionName: 'deployBeaconProxy',
+    args: ['BoardOfDirectors', initData]
+  })
+}
+
+const areAllContractsDeployed = computed(() => {
+  return (
+    props.isBankDeployed &&
+    props.isVotingDeployed &&
+    props.isExpenseDeployed &&
+    props.isExpenseEip712Deployed &&
+    props.isBoDDeployed
+  )
+})
+
+const deployAllContracts = async () => {
+  const currentAddress = useUserDataStore().address as Address
+  const deployments = []
+
+  if (!props.isBankDeployed) {
+    deployments.push({
+      contractType: 'Bank',
+      initializerData: encodeFunctionData({
+        abi: BankABI,
+        functionName: 'initialize',
+        args: [TIPS_ADDRESS, currentAddress]
+      })
+    })
+  }
+
+  if (!props.isVotingDeployed) {
+    deployments.push({
+      contractType: 'Voting',
+      initializerData: encodeFunctionData({
+        abi: VotingABI,
+        functionName: 'initialize',
+        args: [currentAddress]
+      })
+    })
+  }
+
+  if (!props.isExpenseDeployed) {
+    deployments.push({
+      contractType: 'ExpenseAccount',
+      initializerData: encodeFunctionData({
+        abi: ExpenseAccountABI,
+        functionName: 'initialize',
+        args: [currentAddress]
+      })
+    })
+  }
+
+  if (!props.isExpenseEip712Deployed) {
+    deployments.push({
+      contractType: 'ExpenseAccountEIP712',
+      initializerData: encodeFunctionData({
+        abi: ExpenseAccountEIP712ABI,
+        functionName: 'initialize',
+        args: [currentAddress]
+      })
+    })
+  }
+
+  if (!props.isBoDDeployed && props.founders.length > 0) {
+    deployments.push({
+      contractType: 'BoardOfDirectors',
+      initializerData: encodeFunctionData({
+        abi: BoardOfDirectorsABI,
+        functionName: 'initialize',
+        args: [[...props.founders, currentAddress]]
+      })
+    })
+  }
+
+  if (deployments.length > 0) {
+    try {
+      deployAll({
+        address: props.team.officerAddress as Address,
+        abi: OfficerABI,
+        functionName: 'deployAllContracts',
+        args: [deployments]
+      })
+    } catch (error) {
+      log.error(parseError(error))
+      addErrorToast('Error deploying contracts')
+    }
+  }
+}
+
+// Error watchers
+watch(deployBankError, (value) => {
+  if (value) {
+    log.error(parseError(value))
+    addErrorToast('Failed to deploy bank')
+  }
+})
+
+watch(deployVotingError, (value) => {
+  if (value) {
+    addErrorToast('Failed to deploy voting')
+  }
+})
+
+watch(deployExpenseError, (value) => {
+  if (value) {
+    addErrorToast('Failed to deploy expense account')
+  }
+})
+
+watch(deployBoDError, (value) => {
+  if (value) {
+    addErrorToast('Failed to deploy Board of Directors')
+  }
+})
+
+watch(deployAllError, (value) => {
+  if (value) {
+    addErrorToast('Error deploying all contracts')
+  }
+})
+</script>
